requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.40
pydantic>=2.5.0
openai>=1.12.0
python-dotenv>=1.0.0
pydub>=0.25.1

